+++ sdk/core/Azure.Core/src/Identity/Credentials/ManagedIdentityCredential.cs
@@ -142,7 +142,7 @@ namespace Azure.Identity
                 Pipeline = pipeline,
                 PreserveTransport = preserveTransport,
                 Options = options,
-                DisableMtlsProofOfPossession = ResolveDisableMtlsProofOfPossession(options)
+                EnableMtlsProofOfPossession = ResolveEnableMtlsProofOfPossession(options)
             };
         }
 
@@ -159,10 +159,10 @@ namespace Azure.Identity
                 pipeline,
                 preserveTransport);
 
-        private static bool ResolveDisableMtlsProofOfPossession(TokenCredentialOptions options)
+        private static bool ResolveEnableMtlsProofOfPossession(TokenCredentialOptions options)
         {
 #pragma warning disable AZID0004 // Internal usage of experimental mTLS proof-of-possession API
-            return options is ManagedIdentityCredentialOptions managedIdentityCredentialOptions && managedIdentityCredentialOptions.DisableMtlsProofOfPossession;
+            return options is ManagedIdentityCredentialOptions managedIdentityCredentialOptions && managedIdentityCredentialOptions.EnableMtlsProofOfPossession;
 #pragma warning restore AZID0004
         }